+++ src/main/java/net/geforcemods/securitycraft/blockentities/UsernameLoggerBlockEntity.java
@@ -1,5 +1,7 @@
 package net.geforcemods.securitycraft.blockentities;
 
+import java.util.List;
+
 import net.geforcemods.securitycraft.SCContent;
 import net.geforcemods.securitycraft.SecurityCraft;
 import net.geforcemods.securitycraft.api.ILockable;
@@ -40,29 +42,31 @@ public class UsernameLoggerBlockEntity extends DisguisableBlockEntity implements
 			if (cooldown > 0)
 				cooldown--;
 			else if (level.getBestNeighborSignal(worldPosition) > 0) {
-				level.getEntitiesOfClass(PlayerEntity.class, new AxisAlignedBB(worldPosition).inflate(searchRadius.get()), e -> !e.isSpectator()).forEach(this::addPlayer);
-				syncLoggedPlayersToClient();
-				cooldown = TICKS_BETWEEN_ATTACKS;
-			}
-		}
-	}
+				long timestamp = System.currentTimeMillis();
+				List<PlayerEntity> nearbyPlayers = level.getEntitiesOfClass(PlayerEntity.class, new AxisAlignedBB(worldPosition).inflate(searchRadius.get()), e -> !e.isSpectator() && !(isOwnedBy(e) && ignoresOwner() || isAllowed(e)) && !EntityUtils.isInvisible(e) && !wasPlayerRecentlyAdded(e.getName().getString(), timestamp));
 
-	public void addPlayer(PlayerEntity player) {
-		String playerName = player.getName().getString();
-		long timestamp = System.currentTimeMillis();
+				if (!nearbyPlayers.isEmpty()) {
+					boolean changed = false;
+					int playerIndex = 0;
 
-		if (!(isOwnedBy(player) && ignoresOwner()) && !EntityUtils.isInvisible(player) && !wasPlayerRecentlyAdded(playerName, timestamp)) {
-			//ignore players on the allowlist
-			if (isAllowed(player))
-				return;
+					for (int i = 0; i < getPlayers().length && playerIndex < nearbyPlayers.size(); i++) {
+						if (getPlayers()[i] == null || getPlayers()[i].equals("")) {
+							PlayerEntity player = nearbyPlayers.get(playerIndex++);
 
-			for (int i = 0; i < getPlayers().length; i++) {
-				if (getPlayers()[i] == null || getPlayers()[i].equals("")) {
-					getPlayers()[i] = player.getName().getString();
-					getUuids()[i] = player.getGameProfile().getId().toString();
-					getTimestamps()[i] = timestamp;
-					break;
+							getPlayers()[i] = player.getName().getString();
+							getUuids()[i] = player.getGameProfile().getId().toString();
+							getTimestamps()[i] = timestamp;
+							changed = true;
+						}
+					}
+
+					if (changed) {
+						setChanged();
+						syncLoggedPlayersToClient();
+					}
 				}
+
+				cooldown = TICKS_BETWEEN_ATTACKS;
 			}
 		}
 	}