+++ src/main/java/net/geforcemods/securitycraft/blocks/BlockPocketWallBlock.java
@@ -49,14 +49,7 @@ public class BlockPocketWallBlock extends OwnableBlock implements IBlockPocket {
 				if (te1 instanceof BlockPocketBlockEntity) {
 					BlockPocketBlockEntity te = (BlockPocketBlockEntity) te1;
 
-					if (te.getManager() == null)
-						return VoxelShapes.empty();
-
-					if (te.getManager().isAllowed(entity))
-						return VoxelShapes.empty();
-					else if (!te.isOwnedBy((PlayerEntity) entity))
-						return VoxelShapes.block();
-					else
+					if (te.isOwnedBy((PlayerEntity) entity) || te.getManager() == null || te.getManager().isAllowed(entity))
 						return VoxelShapes.empty();
 				}
 			}