+++ src/main/java/net/geforcemods/securitycraft/network/server/ToggleBlockPocketManager.java
@@ -2,11 +2,18 @@ package net.geforcemods.securitycraft.network.server;
 
 import java.util.function.Supplier;
 
+import net.geforcemods.securitycraft.SCContent;
+import net.geforcemods.securitycraft.SecurityCraft;
 import net.geforcemods.securitycraft.blockentities.BlockPocketManagerBlockEntity;
+import net.geforcemods.securitycraft.network.client.BlockPocketManagerFailedActivation;
+import net.geforcemods.securitycraft.util.PlayerUtils;
+import net.geforcemods.securitycraft.util.Utils;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.network.PacketBuffer;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.text.TextFormatting;
+import net.minecraft.util.text.TranslationTextComponent;
 import net.minecraftforge.fml.network.NetworkEvent;
 
 public class ToggleBlockPocketManager {
@@ -16,10 +23,10 @@ public class ToggleBlockPocketManager {
 
 	public ToggleBlockPocketManager() {}
 
-	public ToggleBlockPocketManager(BlockPocketManagerBlockEntity be, boolean enabling, int size) {
+	public ToggleBlockPocketManager(BlockPocketManagerBlockEntity be, boolean enabling) {
 		pos = be.getBlockPos();
 		this.enabling = enabling;
-		this.size = size;
+		size = be.getSize();
 	}
 
 	public ToggleBlockPocketManager(PacketBuffer buf) {
@@ -39,12 +46,21 @@ public class ToggleBlockPocketManager {
 		TileEntity te = player.level.getBlockEntity(pos);
 
 		if (te instanceof BlockPocketManagerBlockEntity && ((BlockPocketManagerBlockEntity) te).isOwnedBy(player)) {
+			TranslationTextComponent feedback;
+
 			((BlockPocketManagerBlockEntity) te).setSize(size);
 
 			if (enabling)
-				((BlockPocketManagerBlockEntity) te).enableMultiblock();
+				feedback = ((BlockPocketManagerBlockEntity) te).enableMultiblock();
 			else
-				((BlockPocketManagerBlockEntity) te).disableMultiblock();
+				feedback = ((BlockPocketManagerBlockEntity) te).disableMultiblock();
+
+			if (feedback != null) {
+				if (enabling && !((BlockPocketManagerBlockEntity) te).isEnabled())
+					SecurityCraft.channel.reply(new BlockPocketManagerFailedActivation(pos), ctx.get());
+
+				PlayerUtils.sendMessageToPlayer(player, Utils.localize(SCContent.BLOCK_POCKET_MANAGER.get().getDescriptionId()), feedback, TextFormatting.DARK_AQUA, false);
+			}
 		}
 	}
 }