+++ Changelog.md
@@ -61,6 +61,7 @@
 - Fix: The Keypad Trapdoor doesn't have any info in JEI
 - Fix: Some recipes do not match vanilla's recipes, allowing for more blocks to be used than intended
 - Removed: "Is always active" option of Keypads and Key Panels. The same behavior can be achieved by setting the "Signal length" option to 0 and activating the block
+- Misc.: Small optimizations to certain network and client performance
 
 --------------------------Changelog for v1.9.8 of SecurityCraft--------------------------